# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RAW.Ninja is an application to store RAW image files.",
        "title": "RAW.Ninja API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/albums/": {
            "get": {
                "description": "Returns a list of albums for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "endpoint for retrieving albums of a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/collection.Resp"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates an album collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint for creating an album collection.",
                "parameters": [
                    {
                        "description": "Data provided for creating the album",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collection.CreateAlbum"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/collection.Resp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
        "/albums/:id": {
            "get": {
                "description": "Returns an album by the ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint fore retrieving a album by ID.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of Collection to retrieve",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collection.Resp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an album by the ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint fore deleting an album by ID.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of Collection to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collection.Resp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            },
            "patch": {
                "description": "Patches an album by the ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint for patching an album by ID.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of Collection to patch",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collection.Resp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
...
        "/photos": {
            "get": {
                "description": "Returns all photo descriptors for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "List user's photo descriptors endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/photo.Response"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.StatusMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "collection.CreateAlbum": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "collection.Resp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/photo.Response"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "common.StatusMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "descriptor.Response": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "filename": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/image.Response"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "type": "string"
                },
                "thumbnail_height": {
                    "type": "integer"
                },
                "thumbnail_width": {
                    "type": "integer"
                },
                "uploaded": {
                    "type": "string"
                }
            }
        },
        "image.PresignedRequest": {
            "type": "object",
            "properties": {
                "method": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "image.Response": {
            "type": "object",
            "properties": {
                "ISO": {
                    "type": "integer"
                },
                "aperture": {
                    "type": "number"
                },
                "camera_make": {
                    "type": "string"
                },
                "camera_model": {
                    "type": "string"
                },
                "camera_sw": {
                    "type": "string"
                },
                "colors": {
                    "type": "integer"
                },
                "data_size": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "lens_make": {
                    "type": "string"
                },
                "lens_model": {
                    "type": "string"
                },
                "shutter": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "photo.Response": {
            "type": "object",
            "properties": {
                "descriptor": {
                    "$ref": "#/definitions/descriptor.Response"
                },
                "id": {
                    "type": "string"
                },
                "raw": {
                    "$ref": "#/definitions/image.PresignedRequest"
                },
                "thumbnail": {
                    "$ref": "#/definitions/image.PresignedRequest"
                }
            }
        },
...
    },
...
}